get_end_points(): avoid a double free on close for the O2MICRO_OZ776 special case.
" In ccid_usb.c you move the extra descriptor from the endpoint 2
descriptor to the interface descriptor. The problem is, when libusb try
to free the descriptors associated to this device on shutdown or if the
device disappears, it tries to free the same buffer twice. For me it
means the pcscd is killed with a glibc detected "double free". "

Thanks to Michael Wegner for the patch


git-svn-id: svn://anonscm.debian.org/pcsclite/trunk@2510 0ce88b0d-b2fd-0310-8134-9614164e65ea