main(): do not fail if pcscd.comm is already present but pcscd.pid is not present.
This is the case if pcscd has been configured in systemd but has been
stoped. When launched by hand (for a debug session for example) the
pcscd.comm file is still present but not used anymore.

The pcscd.comm file will be removed when pcscd manually started will
exit but the file will be recreated by "systemctl start pcscd.service"


git-svn-id: svn://anonscm.debian.org/pcsclite/trunk@6072 0ce88b0d-b2fd-0310-8134-9614164e65ea