Log to stdout by default
This log service is only used on Mac OS X. Using stderr is problematic
when other log mesages are using stdout. The order is not respected if
stdout and stderr are redirected to a file. For example when using
handler_test.

Now the logs are sent to stdout (by default) and also flushed after each
line. It is possible to use tee(1).


git-svn-id: svn://anonscm.debian.org/pcsclite/trunk@6179 0ce88b0d-b2fd-0310-8134-9614164e65ea