Use LONG/ULONG instead of long/unsigned long
The problem is that PCSC on Mac OS X do not use long but uint32_t. So a
(LONG * or LPDWORD *) parameter (like pcchReaders in SCardListReaders)
is not the same on Linux and Mac OS X for 64-bits CPUs.

LONG is correctly defined on both Linux and Mac OS X so we use it
instead.


git-svn-id: svn://anonscm.debian.org/pcsclite/trunk@6195 0ce88b0d-b2fd-0310-8134-9614164e65ea