Add support of Gemalto firmware features
Gemalto readers support an Escape command to report features supported
(or not) by the reader. The reader can also report if specific bugs are
fixed or not.

One of the fixed bug (in recent readers) is the bNumberMessage
management used in the Secure Modify CCID command.

Thanks to Jean Lusetti for the patch


git-svn-id: svn://anonscm.debian.org/pcsclite/trunk@6252 0ce88b0d-b2fd-0310-8134-9614164e65ea