Use auto suspend for CCID devices only
The rule used "*:0b0000:*" and so was used even for composite devices
using a CCID interface. It was problematic for a keybord + CCID
interface since the keyboard was also automatically systended after a
few seconds of inactivity.

The new rule uses ":0b0000:" so composite devices will not be configured
for USB automatic suspend.

Closes Alioth bug [#313445] "Do not activate USB suspend for composite
devices: keyboard"
https://alioth.debian.org/tracker/index.php?func=detail&aid=313445&group_id=30105&atid=410085


git-svn-id: svn://anonscm.debian.org/pcsclite/trunk@6275 0ce88b0d-b2fd-0310-8134-9614164e65ea