t1_transceive(): avoids a infinite retry loop
Add a check for ISO 7816-3 Rule 7.4.2 (2 retried before resync) in case
of "wrong nr".

" currently I have an endless loop in t1_transceive.  It happens from
time to time with the aladdin token.

I added some printfs to the code.
I can see, that the errors are detected and then a t1_build
is called with a T1_OTHER_ERROR, retries are decreased,
but this is never evaluated in the loop path
and the retries counter counts and counts....
I stopped it at: retries -1796. "

Thanks to Martin Vogt for the bug report and patch
http://archives.neohapsis.com/archives/dev/muscle/2012-q2/0076.html


git-svn-id: svn://anonscm.debian.org/pcsclite/trunk@6327 0ce88b0d-b2fd-0310-8134-9614164e65ea